Add files via upload
"Dataset for Car Pricing Prediction"

This dataset contains various features of cars collected from the American market, such as engine specifications, car make, model, year, and other attributes that influence pricing. The dataset is used to train machine learning models for predicting car prices. It includes both numerical and categorical variables, and the target variable is the price of the car. The dataset has been preprocessed, including handling missing values and feature scaling, to make it suitable for regression analysis.